+++ src/pages/company/CompanyList.js
@@ -24,6 +24,7 @@ import { useForm } from "react-hook-form";
 import { CompanyContext } from "./CompanyContext";
 import CompanyServices from "../../services/CompanyServices";
 import CountriesServices from "../../services/CountriesServices";
+import swal from "sweetalert";
 
 const CompanyList = () => {
   const { contextData } = useContext(CompanyContext);
@@ -238,10 +239,35 @@ const CompanyList = () => {
 
   // Function to change to delete property for an item
   const deleteUser = async (id) => {
+    // await CompanyServices.deleteCompany(id);
+    // getCompany();
     try {
-      await CompanyServices.deleteCompany(id);
-      getCompany();
-    } catch (error) {}
+      await swal({
+        title: "¿Estás seguro?",
+        text: "Se eliminará el Socio Estrategico seleccionado!",
+        icon: "warning",
+        dangerMode: true,
+        buttons: {
+          confirm: { text: "Aceptar", className: "bg-primary" },
+          cancel: "Cancelar",
+        },
+      }).then((resDelete) => {
+        if (resDelete) {
+          getCompany();
+          swal("Listo! Socio Estratégico eliminado exitosamente!", {
+            icon: "success",
+            timer: "2000",
+            buttons: {
+              confirm: { text: "Listo", className: "bg-primary" },
+            },
+          });
+          CompanyServices.deleteCompany(id);
+          getCompany();
+        }
+      });
+    } catch (error) {
+      throw new Error("Error deleting record!");
+    }
   };
 
   // Get current list, pagination
@@ -268,7 +294,7 @@ const CompanyList = () => {
           <BlockBetween>
             <BlockHeadContent>
               <BlockTitle tag="h3" page>
-                Lista de Empresa
+                Lista de Socios Estratégicos
               </BlockTitle>
               <BlockDes className="text-soft">
                 <p>Total {data.length} Empresa</p>
@@ -290,7 +316,7 @@ const CompanyList = () => {
                     <li className="nk-block-tools-opt">
                       <Button color="primary" onClick={() => setModal({ add: true })}>
                         <Icon name="plus" className="mr-1"></Icon>
-                        Agregar Empresa
+                        Agregar Socio Estratégico
                       </Button>
                     </li>
                   </ul>
@@ -317,10 +343,10 @@ const CompanyList = () => {
                   <span className="sub-text">RUT</span>
                 </DataTableRow>
                 <DataTableRow>
-                  <span className="sub-text">Telefono</span>
+                  <span className="sub-text">Teléfono</span>
                 </DataTableRow>
                 <DataTableRow>
-                  <span className="sub-text"></span>
+                  <span className="sub-text">Acción</span>
                 </DataTableRow>
               </DataTableHead>
               {/*Head*/}
@@ -344,24 +370,24 @@ const CompanyList = () => {
                       </DataTableRow>
                       <DataTableRow className="nk-tb-col-tools">
                         <ul className="nk-tb-actions gx-1">
-                          <li className="nk-tb-action-hidden" onClick={() => onEditClick(item.id, item)}>
+                          <li className="nk-tb-action" onClick={() => onEditClick(item.id, item)}>
                             <TooltipComponent
                               tag="a"
                               containerClassName="btn btn-trigger btn-icon"
                               id={"edit" + 1}
                               icon="edit-alt-fill"
                               direction="top"
-                              text="Edit"
+                              text="Editar"
                             />
                           </li>
-                          <li className="nk-tb-action-hidden" onClick={() => deleteUser(item.id)}>
+                          <li className="nk-tb-action" onClick={() => deleteUser(item.id)}>
                             <TooltipComponent
                               tag="a"
                               containerClassName="btn btn-trigger btn-icon"
                               id={"delete" + 1}
                               icon="trash-fill"
                               direction="top"
-                              text="Delete"
+                              text="Eliminar"
                             />
                           </li>
                         </ul>
@@ -370,25 +396,24 @@ const CompanyList = () => {
                   ))
                 : null}
             </div>
-
-            <PreviewAltCard>
-              {currentItems.length > 0 ? (
-                <PaginationComponent
-                  itemPerPage={itemPerPage}
-                  totalItems={data.length}
-                  paginate={paginate}
-                  currentPage={currentPage}
-                />
-              ) : (
-                <div className="text-center">
-                  <span className="text-silent">Sin Registros</span>
-                </div>
-              )}
-            </PreviewAltCard>
           </div>
+          <PreviewAltCard>
+            {currentItems.length > 0 ? (
+              <PaginationComponent
+                itemPerPage={itemPerPage}
+                totalItems={data.length}
+                paginate={paginate}
+                currentPage={currentPage}
+              />
+            ) : (
+              <div className="text-center">
+                <span className="text-silent">Sin Registros</span>
+              </div>
+            )}
+          </PreviewAltCard>
         </Block>
 
-        <Modal isOpen={modal.add} toggle={() => setModal({ add: false })} className="modal-dialog-centered" size="lg">
+        <Modal isOpen={modal.add} toggle={() => setModal({ add: true })} className="modal-dialog-centered" size="lg">
           <ModalBody>
             <a
               href="#close"
@@ -585,7 +610,7 @@ const CompanyList = () => {
                       <ul className="align-center flex-wrap flex-sm-nowrap gx-4 gy-2">
                         <li>
                           <Button color="primary" size="md" type="submit">
-                            Agregar Empresa
+                            Agregar Socio Estratégico
                           </Button>
                         </li>
                         <li>
@@ -610,7 +635,7 @@ const CompanyList = () => {
           </ModalBody>
         </Modal>
 
-        <Modal isOpen={modal.edit} toggle={() => setModal({ edit: false })} className="modal-dialog-centered" size="lg">
+        <Modal isOpen={modal.edit} toggle={() => setModal({ edit: true })} className="modal-dialog-centered" size="lg">
           <ModalBody>
             <a
               href="#cancel"
