+++ src/pages/documents/DocumentsList.js
@@ -21,6 +21,7 @@ import Content from "../../layout/content/Content";
 import Head from "../../layout/head/Head";
 import { useForm } from "react-hook-form";
 import DocumentsServices from "../../services/DocumentsServices";
+import swal from "sweetalert";
 
 const DocumentsList = () => {
   const [data, setData] = useState([]);
@@ -110,10 +111,35 @@ const DocumentsList = () => {
 
   // Function to change to delete property for an item
   const deleteDocument = async (id) => {
+    // await DocumentsServices.deleteDocument(id);
+    // getDocuments();
     try {
-      await DocumentsServices.deleteDocument(id);
-      getDocuments();
-    } catch (error) {}
+      await swal({
+        title: "¿Estás seguro?",
+        text: "Se eliminará el Documento seleccionado!",
+        icon: "warning",
+        dangerMode: true,
+        buttons: {
+          confirm: { text: "Aceptar", className: "bg-primary" },
+          cancel: "Cancelar",
+        },
+      }).then((resDelete) => {
+        if (resDelete) {
+          getDocuments();
+          swal("Listo! Documento eliminado exitosamente!", {
+            icon: "success",
+            timer: "2000",
+            buttons: {
+              confirm: { text: "Listo", className: "bg-primary" },
+            },
+          });
+          DocumentsServices.deleteDocument(id);
+          getDocuments();
+        }
+      });
+    } catch (error) {
+      throw new Error("Error deleting record!");
+    }
   };
 
   // Get current list, pagination
@@ -174,7 +200,7 @@ const DocumentsList = () => {
             <div className="nk-tb-list is-separate is-medium mb-3">
               <DataTableHead className="nk-tb-item">
                 <DataTableRow className="text-center">
-                  <span className="sub-text">N. Documento</span>
+                  <span className="sub-text">Documento</span>
                 </DataTableRow>
                 <DataTableRow className="text-center">
                   <span className="sub-text">Nombre</span>
@@ -227,25 +253,24 @@ const DocumentsList = () => {
                   ))
                 : null}
             </div>
-
-            <PreviewAltCard>
-              {currentItems.length > 0 ? (
-                <PaginationComponent
-                  itemPerPage={itemPerPage}
-                  totalItems={currentItems.length}
-                  paginate={paginate}
-                  currentPage={currentPage}
-                />
-              ) : (
-                <div className="text-center">
-                  <span className="text-silent">Sin Registros</span>
-                </div>
-              )}
-            </PreviewAltCard>
           </div>
+          <PreviewAltCard>
+            {currentItems.length > 0 ? (
+              <PaginationComponent
+                itemPerPage={itemPerPage}
+                totalItems={currentItems.length}
+                paginate={paginate}
+                currentPage={currentPage}
+              />
+            ) : (
+              <div className="text-center">
+                <span className="text-silent">Sin Registros</span>
+              </div>
+            )}
+          </PreviewAltCard>
         </Block>
 
-        <Modal isOpen={modal.add} toggle={() => setModal({ add: false })} className="modal-dialog-centered" size="lg">
+        <Modal isOpen={modal.add} toggle={() => setModal({ add: true })} className="modal-dialog-centered" size="lg">
           <ModalBody>
             <a
               href="#close"
@@ -327,7 +352,7 @@ const DocumentsList = () => {
           </ModalBody>
         </Modal>
 
-        <Modal isOpen={modal.edit} toggle={() => setModal({ edit: false })} className="modal-dialog-centered" size="lg">
+        <Modal isOpen={modal.edit} toggle={() => setModal({ edit: true })} className="modal-dialog-centered" size="lg">
           <ModalBody>
             <a
               href="#cancel"
